Refactors data grabbing code
Before, we handled temporal gaps in the data on the server-side, but
since this was only useful for cubism.js, we now provide the timestamps
with the readings so that they are more universally useful.

The gaps in the data are now dealt with using in the datasource which
drives cubism.js and the historical graph has been updated so that it
no longer has to calculate datetime values.

As such the API has been changed - the returned JS array of objects with
the following keys:
 - ts => the timestammp of the datapoint
 - temperature => the temperature reading
 - valuye => (deprecated) the temperature reading.

This is to allow for new time-series readings which should be appearing
soon.